pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aws-tui"
version = "0.1.0"
description = "Textual TUI for browsing EC2 instances and starting SSM sessions."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.34.0",
  "pyyaml>=6.0",
  "textual>=0.58.0",
]

[project.scripts]
aws-tui = "aws_tui.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aws_tui"]

[tool.hatch.build.targets.sdist]
include = [
  "/src/aws_tui",
  "/port-forwards.yaml",
  "/README.md",
  "/pyproject.toml",
]
